Fix warnings and convert to XCTest

+++ RZCollectionList-Demo/RZCollectionList-Demo/Classes/ViewControllers/ArrayListViewController.m
@@ -69,7 +69,7 @@
 {
     static NSUInteger totalCount = 0;
     ++totalCount;
-    [self.arrayList addObject:[ListItemObject listItemObjectWithName:[NSString stringWithFormat:@"Item %u", totalCount] subtitle:nil] toSection:0];
+    [self.arrayList addObject:[ListItemObject listItemObjectWithName:[NSString stringWithFormat:@"Item %lu", (unsigned long)totalCount] subtitle:nil] toSection:0];
 }
 
 #pragma mark - UITableViewDelegate
